requirements.txt
streamlit==1.28.0
python-dotenv
PyPDF2==3.0.1
langchain==0.0.314